Create libsmtutil
