+++ libsmtutil/CVC4Interface.h
@@ -17,52 +17,54 @@
 
 #pragma once
 
-#include <libsolidity/formal/SolverInterface.h>
+#include <libsmtutil/SolverInterface.h>
 #include <boost/noncopyable.hpp>
-#include <z3++.h>
+
+#if defined(__GLIBC__)
+// The CVC4 headers includes the deprecated system headers <ext/hash_map>
+// and <ext/hash_set>. These headers cause a warning that will break the
+// build, unless _GLIBCXX_PERMIT_BACKWARD_HASH is set.
+#define _GLIBCXX_PERMIT_BACKWARD_HASH
+#endif
+
+#include <cvc4/cvc4.h>
+
+#if defined(__GLIBC__)
+#undef _GLIBCXX_PERMIT_BACKWARD_HASH
+#endif
 
 namespace solidity::frontend::smt
 {
 
-class Z3Interface: public SolverInterface, public boost::noncopyable
+class CVC4Interface: public SolverInterface, public boost::noncopyable
 {
 public:
-	Z3Interface();
+	CVC4Interface();
 
 	void reset() override;
 
 	void push() override;
 	void pop() override;
 
-	void declareVariable(std::string const& _name, SortPointer const& _sort) override;
+	void declareVariable(std::string const&, SortPointer const&) override;
 
 	void addAssertion(Expression const& _expr) override;
 	std::pair<CheckResult, std::vector<std::string>> check(std::vector<Expression> const& _expressionsToEvaluate) override;
 
-	z3::expr toZ3Expr(Expression const& _expr);
-
-	std::map<std::string, z3::expr> constants() const { return m_constants; }
-	std::map<std::string, z3::func_decl> functions() const { return m_functions; }
-
-	z3::context* context() { return &m_context; }
-
-	// Z3 "basic resources" limit.
-	// This is used to make the runs more deterministic and platform/machine independent.
-	// The tests start failing for Z3 with less than 20000000,
-	// so using double that.
-	static int const resourceLimit = 40000000;
-
 private:
-	void declareFunction(std::string const& _name, Sort const& _sort);
+	CVC4::Expr toCVC4Expr(Expression const& _expr);
+	CVC4::Type cvc4Sort(smt::Sort const& _sort);
+	std::vector<CVC4::Type> cvc4Sort(std::vector<smt::SortPointer> const& _sorts);
 
-	z3::sort z3Sort(smt::Sort const& _sort);
-	z3::sort_vector z3Sort(std::vector<smt::SortPointer> const& _sorts);
+	CVC4::ExprManager m_context;
+	CVC4::SmtEngine m_solver;
+	std::map<std::string, CVC4::Expr> m_variables;
 
-	z3::context m_context;
-	z3::solver m_solver;
-
-	std::map<std::string, z3::expr> m_constants;
-	std::map<std::string, z3::func_decl> m_functions;
+	// CVC4 "basic resources" limit.
+	// This is used to make the runs more deterministic and platform/machine independent.
+	// The tests start failing for CVC4 with less than 6000,
+	// so using double that.
+	static int const resourceLimit = 12000;
 };
 
 }
