+++ libsmtutil/SMTPortfolio.h
@@ -18,7 +18,7 @@
 #pragma once
 
 
-#include <libsolidity/formal/SolverInterface.h>
+#include <libsmtutil/SolverInterface.h>
 #include <libsolidity/interface/ReadFile.h>
 #include <libsolutil/FixedHash.h>
